use linux.rm_dir_force instead of rm -rf

See merge request zstackio/zstack-utility!2502

+++ zstacklib/zstacklib/utils/linux.py
@@ -88,9 +88,15 @@ def rm_file_force(fpath):
     except:
         pass
 
-def rm_dir_force(dpath):
-    if os.path.exists(dpath):
+black_dpath_list = ["", "/", "*", "/root", "/var", "/bin", "/lib", "/sys"]
+
+def rm_dir_force(dpath, only_check=False):
+    if dpath.strip() in black_dpath_list:
+        raise Exception("how dare you delete directory %s" % dpath)
+    if os.path.exists(dpath) and not only_check:
         shutil.rmtree(dpath)
+    else:
+        return dpath
 
 def rm_file_checked(fpath):
     if not os.path.exists(fpath):
